refactored config into dags
