+++ dags/sample.env
@@ -1,0 +1,25 @@
+CONFIG_PATH = /Users/ryanlattanzi/projects/algo_trading/dags/config/config.yml
+DATA_BUCKET=algo-trading-price-data
+DATA_KEY={ticker}/{run_date}.csv
+LOG_BUCKET=algo-trading-logs
+LOG_KEY={log_name}/{run_date}.log
+
+POSTGRES_USER=dev
+POSTGRES_PASSWORD=password
+POSTGRES_DB=price_db
+POSTGRES_HOST=localhost
+POSTGRES_PORT=5432
+PGDATA=/var/lib/postgresql/data/pgdata
+
+REDIS_HOST=localhost
+REDIS_PORT=6379
+REDIS_DB=0
+REDIS_PASSWORD=redispassword
+
+MINIO_ENDPOINT=http://localhost:9000
+MINIO_REGION=us-east-1
+MINIO_ROOT_USER=admin
+MINIO_ROOT_PASSWORD=password
+
+PGADMIN_DEFAULT_EMAIL=admin@admin.com
+PGADMIN_DEFAULT_PASSWORD=password
