+++ email_service/sample.env
@@ -1,10 +1,12 @@
 EMAIL_MANAGER_STORE=redis
-EMAIL_MANAGER_HOST=localhost
+EMAIL_MANAGER_HOST=algo_trading-redis-1
 EMAIL_MANAGER_PORT=6379
 EMAIL_MANAGER_DB=1
 EMAIL_MANAGER_PASSWORD=redispassword
 
 SMTP_SERVER=smtp.gmail.com
 SSL_PORT=465
-SENDER_EMAIL=example@gmail.com
-SENDER_EMAIL_PASSWORD=password
+SENDER_EMAIL=algotrading.producer@gmail.com
+SENDER_EMAIL_PASSWORD=Producer#1
+
+LOG_LEVEL=info
